Remove redundant message references
When Bot._onMessage() executes, it now stores that message in a
variable, which is utilized by all commands globally.

All setMsg*() functions have been converted to sendMsg*(); this
allows commands to effortlessly send messages to the last
message channel.

The caveat with this change is that if any async operations fail to
return before another Bot._onMessage() event is executed, we
risk sending a message to the wrong channel.