cmd(qstn): Add `doc` sub-cmd
Enables converting an existing item back into an Item Document.
This can be useful in cases where the original user-created
document, has been lost or misplaced.

The document will not necessarily be a 1 to 1 representation of
the original, because it can only decode the questions back into
their short form. That being said, it is functionally equivalent to
the original document.